requirements.txt
flask~=2.3.2
sqlalchemy~=2.0.19
flask_sqlalchemy
flask_bcrypt
flask_resize
flask_wtf
Flask-QRcode
flask_login
Flask-mail
bleach~=6.0.0
pdfkit~=1.0.0
python-slugify
MySQLClient
python-dotenv
Pillow~=10.0.0
